requirements.txt
streamlit>=1.38,<2
requests>=2.32,<3
python-dotenv>=1.0,<2
openai>=1.50,<2
